64-bit conversion changes
git-svn-id: svn://orca.physics.unc.edu/Orca64@11899 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/Halo/HaloModel.h
@@ -49,8 +49,8 @@
 - (void) takeOverRunning;
 - (NSMutableArray*) emailList;
 - (void) setEmailList:(NSMutableArray*)aEmailList;
-- (void) addAddress:(id)anAddress atIndex:(int)anIndex;
-- (void) removeAddressAtIndex:(int) anIndex;
+- (void) addAddress:(id)anAddress atIndex:(NSUInteger)anIndex;
+- (void) removeAddressAtIndex:(NSUInteger) anIndex;
 - (int) heartBeatIndex;
 - (void) setHeartBeatIndex:(int)aHeartBeatIndex;
 - (int) heartbeatSeconds;
