+++ Source/Main Dialogs/CommandCenter/Scripting/ORScriptIDEController.m
@@ -729,8 +729,8 @@
 - (IBAction) removeInput:(id)sender
 {
 	NSIndexSet* indexSet = [inputVariablesTableView selectedRowIndexes];
-	int i;
-	int last = (int)[indexSet lastIndex];
+	NSUInteger i;
+	NSUInteger last = (int)[indexSet lastIndex];
 	for(i=last;i!=NSNotFound;i = [indexSet indexLessThanIndex:i]){
 		[model removeInputValue:i];
 	}
